requirements.txt
dj-database-url==0.4.2
Django==2.2.5
gunicorn==19.7.1
psycopg2==2.8.3
whitenoise==3.3.0
djangorestframework==3.10.3
